requirements.txt
aiohttp==3.8.1
pyyaml==6.0
feedgen==0.9.0
python-dotenv==0.19.0
requests==2.31.0
asyncio==3.4.3
aiofiles==23.2.1
loguru==0.7.2
pendulum==2.1.2
emoji==2.9.0
retrying==1.3.4
